+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/parsers/antlr/statements/ms/GrantMsPrivilege.java
@@ -20,6 +20,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 import java.util.Map.Entry;
 
@@ -77,7 +78,9 @@ public class GrantMsPrivilege extends MsParserAbstract {
         }
 
         List<String> roles = ctx.name_list().id().stream()
-                .map(ParserRuleContext::getText).map(MsDiffUtils::quoteName).toList();
+                .map(ParserRuleContext::getText)
+                .map(MsDiffUtils::quoteName)
+                .toList();
 
         Columns_permissionsContext columnsCtx = ctx.columns_permissions();
         if (columnsCtx != null) {
@@ -86,7 +89,9 @@ public class GrantMsPrivilege extends MsParserAbstract {
         }
 
         List<String> permissions = ctx.permissions().permission().stream()
-                .map(ParserAbstract::getFullCtxText).toList();
+                .map(ParserAbstract::getFullCtxText)
+                .map(e -> e.toUpperCase(Locale.ROOT))
+                .toList();
 
         PgStatement st = getStatement(nameCtx);
 
@@ -160,7 +165,7 @@ public class GrantMsPrivilege extends MsParserAbstract {
         // collect information about column privileges
         Map<String, Entry<IdContext, List<String>>> colPriv = new HashMap<>();
         for (Table_column_privilegesContext priv : columnsCtx.table_column_privileges()) {
-            String privName = getFullCtxText(priv.permission());
+            String privName = getFullCtxText(priv.permission()).toUpperCase(Locale.ROOT);
             for (IdContext col : priv.name_list_in_brackets().id()) {
                 colPriv.computeIfAbsent(col.getText(),
                         k -> new SimpleEntry<>(col, new ArrayList<>())).getValue().add(privName);