+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fixed error when explicitly specifying default schema in dump.
 - Fixed parser rule for Greenplum `CREATE TABLE AS`.
+- Fixed a bug when comparing partitions in Greenplum.
 
 ### Removed